Refactor torch into seperate component
Fixes issues with state updates breaking the camera
