+++ src/Components/QrScanner/QrScannerPlugin.tsx
@@ -1,8 +1,9 @@
 // file = QrScannerPlugin.jsx
-import {MutableRefObject, useEffect, useRef, useState} from 'react';
-import {ArrowUpTrayIcon, BoltIcon, BoltSlashIcon} from '@heroicons/react/24/solid';
+import {MutableRefObject, useEffect, useRef} from 'react';
+import {ArrowUpTrayIcon} from '@heroicons/react/24/solid';
 import {Html5Qrcode, Html5QrcodeScannerState, Html5QrcodeSupportedFormats} from 'html5-qrcode';
 import {checkCameraPermissions} from '../../utils/media';
+import {TorchButton} from './TorchButton';
 import classes from './QrScanner.module.css';
 
 // Id of the HTML element used by the Html5QrcodeScanner.
@@ -51,8 +52,6 @@ export default function QrScannerPlugin({
 }: QrProps) {
   const aspectRatio = calcAspectRatio();
   const html5CustomScanner: MutableRefObject<Html5Qrcode | null> = useRef(null);
-  const torchStateRef = useRef(false);
-  const [, setRender] = useState(false); // TODO: Find a better way to force a re-render without using the state (this causes the camera to reload)
 
   useEffect(() => {
     const showQRCode = async () => {
@@ -105,26 +104,13 @@ export default function QrScannerPlugin({
     onPermRefused,
   ]);
 
-  const toggleTorch = async () => {
-    try {
-      const track = html5CustomScanner.current?.getRunningTrackCameraCapabilities();
-      if (track && track.torchFeature().isSupported()) {
-        torchStateRef.current = !torchStateRef.current;
-        await track.torchFeature().apply(torchStateRef.current);
-        setRender(prev => !prev); // TODO: Hacky
-      }
-    } catch (error) {
-      console.warn('Error toggling torch:', error);
-    }
-  };
-
   return (
     <>
       <div className={classes.wrapper}>
         <ShadedRegion size={qrbox}></ShadedRegion>
         <div id={qrcodeRegionId} />
       </div>
-      <ToggleTorchButton onClick={toggleTorch} toggled={!torchStateRef.current} />
+      <TorchButton html5CustomScanner={html5CustomScanner} />
     </>
   );
 }
@@ -151,21 +137,6 @@ export function FileUploadScanner({
         <span>Upload QR code image</span>
       </label>
       <div id={qrcodeFileRegionId}></div>
-    </div>
-  );
-}
-
-export function ToggleTorchButton({onClick, toggled}: {onClick: () => void; toggled: boolean}) {
-  // TODO: Hide away the button when the torch is not supported
-  return (
-    <div className="mt-4 flex justify-center">
-      <div onClick={onClick} className="inline-flex rounded-full bg-primary text-white">
-        {toggled ? (
-          <BoltIcon className="mx-2 my-2 h-16 w-16" />
-        ) : (
-          <BoltSlashIcon className="mx-2 my-2 h-16 w-16" />
-        )}
-      </div>
     </div>
   );
 }
